pyproject.toml
[project]
name = "aichatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gtts>=2.5.4",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.1.5",
    "langdetect>=1.0.9",
    "langgraph>=0.4.8",
    "openai>=1.90.0",
    "playsound>=1.3.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.0",
    "speechrecognition>=3.14.3",
]
